% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{abrv2rec_type}
\alias{abrv2rec_type}
\title{Convert abreviated fhx file event char to rec_type char.}
\usage{
abrv2rec_type(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string.
}
\description{
Convert abreviated fhx file event char to rec_type char.
}
