% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brn_census.R
\name{brn_census}
\alias{brn_census}
\title{Create a Burner Response Nodule (BRN) object}
\usage{
brn_census(dataLocation, years, qid, weights, metaData)
}
\arguments{
\item{dataLocation}{file path of the folder containing the Census survey data}

\item{years}{numeric vector with the same length as filePaths of years being
read into the census.data object and corresponding to each file path.}

\item{qid}{character or numeric vector of either survey question id names or
numbers present in the metaData object}

\item{weights}{character vector with the same length as year corresponding
to the name of the weight variable used for each year.}

\item{metaData}{A census.meta object.  Only questions included in the
census.meta object will be included in the census.data object.}
}
\value{
A census.data object.
}
\description{
Create a Burner Response Nodule (BRN) object
}
\examples{
# example code
}
