% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brn_table.R
\name{brn_table}
\alias{brn_table}
\title{Table Creation and Cross Tabulation for Weighted Census estimates}
\usage{
brn_table(
  brnObject,
  formula,
  years,
  type = c("proportion", "populationCounts")
)
}
\arguments{
\item{formula}{A one- or two-sided formula.  See details.}

\item{years}{A numeric vector of years to include in the table.}

\item{type}{One of "proportion" or "populationCounts".  See details.}

\item{census}{A brn object.}
}
\description{
Table Creation and Cross Tabulation for Weighted Census estimates
}
\details{
TODO
}
