% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check for duplicate observations in an fhx object.}
\usage{
check_duplicates(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
A \code{x} or stop() is thrown.
}
\description{
Check for duplicate observations in an fhx object.
}
\examples{
data(lgr2)
data(pgm)
burnr:::check_duplicates(lgr2 + pgm)

}
