% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{check_series}
\alias{check_series}
\title{Diagnostic checks of common data issues in FHX objects}
\usage{
check_series(x, keep_checks = FALSE)
}
\arguments{
\item{x}{An \code{fhx} object.}

\item{keep_checks}{Boolean indicating whether the check data should be
provided for further diagnostics. Defaults to FALSE, and only warnings or
information are provided.}
}
\value{
If necessary, a set of warnings and informational text to indicate
where and when issues are found. If \code{keep_checks = TRUE}, a list object
with the set of check data.
}
\description{
Diagnostic checks of common data issues in FHX objects
}
\examples{
# Nothing is returned unless an issue is found
check_series(lgr2)

# Informative messages flag potential issues
check_series(pgm)

# Danger flags warn of data quality and usability issues
\dontrun{
d <- read_fhx(paste0("https://github.com/ltrr-arizona-edu/burnr/",
                      "files/818119/Dumb_broke_fhx.txt"))
check_series(d)
}

}
