% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cman_2men.R
\name{cman.2men}
\alias{cman.2men}
\title{Concatenate two man objects}
\usage{
cman.2men(man1, man2)
}
\arguments{
\item{man1}{A 'man' object to concatenate.}

\item{man2}{A 'man' object to concatenate.}
}
\value{
A 'man' object.
}
\description{
Concatenate two man objects
}
