% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cman_cyears.R
\name{cman.cyears}
\alias{cman.cyears}
\title{Combine Years of a 'qname' Present in Two 'man' Objects}
\usage{
cman.cyears(qnm, mn1, mn2)
}
\arguments{
\item{qnm}{A qname value for a Census question. This qname should be present
in both mn1 and mn2.}

\item{mn1}{A 'man' object to concatenate.}

\item{mn2}{A 'man' object to concatenate.}
}
\value{
A list that can be converted to a 'man' object.
}
\description{
Combine Years of a 'qname' Present in Two 'man' Objects
}
