% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine two fhx objects.}
\usage{
combine(a, b)
}
\arguments{
\item{a}{An fhx instance.}

\item{b}{The fhx instance to be appended.}
}
\value{
An fhx instance with the information from \code{a} and \code{b}.
}
\description{
Combine two fhx objects.
}

