% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_injury}
\alias{count_injury}
\title{Number of injury events in an fhx series.}
\usage{
count_injury(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
The number of injury events observed in the series.
}
\description{
Number of injury events in an fhx series.
}
