% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_recording}
\alias{count_recording}
\title{Number of recording years in an fhx series.}
\usage{
count_recording(x, injury_event = FALSE)
}
\arguments{
\item{x}{An fhx object.}

\item{injury_event}{Boolean indicating whether injuries should be considered event.}
}
\value{
The number of recording events observed in the series.
}
\description{
Number of recording years in an fhx series.
}

