% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_scar}
\alias{count_scar}
\title{Number of scar events in an \code{fhx} object}
\usage{
count_scar(x)
}
\arguments{
\item{x}{An \code{fhx} object.}
}
\value{
The number of fire scar events in \code{x}
}
\description{
Number of scar events in an \code{fhx} object
}
\seealso{
\itemize{
\item \code{\link[=count_injury]{count_injury()}} Count the injuries in an \code{fhx} object.
\item \code{\link[=series_stats]{series_stats()}} basic statistics for series in an \code{fhx} object.
}
}
