% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_year_span}
\alias{count_year_span}
\title{Number of years of an fhx series.}
\usage{
count_year_span(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
The difference between the first and last observations in the series. 'NA' will be returned if 'NA' is in 'x$year'.
}
\description{
Number of years of an fhx series.
}
