% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{delete}
\alias{delete}
\title{Remove series or year(s) from an fhx object and return.}
\usage{
delete(x, s, yr)
}
\arguments{
\item{x}{An fhx object.}

\item{s}{Character vector of series to erase from x.}

\item{yr}{Integer vector of years to erase from x.}
}
\value{
An fhx object with observations erased.
}
\description{
Remove series or year(s) from an fhx object and return.
}
\details{
You can combine s and yr to specify years within select series to remove.
}

