% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fhx}
\alias{fhx}
\title{Constructor for S3 fhx class.}
\usage{
fhx(year, series, rec_type)
}
\arguments{
\item{year}{A numeric vector of observation years for each \code{series} and \code{rec_type} argument.}

\item{series}{A factor of series names for each \code{year} and \code{rec_type} argument.}

\item{rec_type}{A factor of ring types for each element in \code{year} and \code{series}.}
}
\value{
An fhx instance.
}
\description{
Constructor for S3 fhx class.
}
