% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_recording}
\alias{find_recording}
\title{Subset `rings` data.frame to years that are considered recording.}
\usage{
find_recording(x, injury_event)
}
\arguments{
\item{x}{A an fhx object dataframe.}

\item{injury_event}{Boolean indicating whether injuries should be considered event.}
}
\value{
A dataframe with a column of each year which is 'recording'.
}
\description{
Subset `rings` data.frame to years that are considered recording.
}
\examples{
require(plyr)
data(lgr2)
ddply(lgr2$rings, 'series', burnr:::find_recording, injury_event = TRUE)

}

