% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{first_year}
\alias{first_year}
\title{First (earliest) year of an fhx series.}
\usage{
first_year(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
The minimum or first year of series in 'x'.
}
\description{
First (earliest) year of an fhx series.
}

