% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{fit_distr}
\alias{fit_distr}
\title{Fit a theoretical distribution to intervals.}
\usage{
fit_distr(intrv, distr, ...)
}
\arguments{
\item{intrv}{An intervals instance.}

\item{distr}{String giving desired distribution to fit. Can be "weibull" or "lognormal". Default is "weibull".}

\item{...}{Arguments passed to \code{MASS::fitdistr}.}
}
\value{
An intervals instance with fit distribution.
}
\description{
Fit a theoretical distribution to intervals.
}
