% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_ggplot}
\alias{get_ggplot}
\title{Create an ggplot2 object for plotting.}
\usage{
get_ggplot(x, spp, sppid, cluster, clusterid, facet_type = "grid",
  ylabels = TRUE, yearlims = FALSE, composite_rug = FALSE,
  filter_prop = 0.25, filter_min = 2, legend = FALSE,
  event_size = c(Scar = 4, Injury = 2, `Pith/Bark` = 1.5),
  rugbuffer_size = 2, rugdivide_pos = 2)
}
\arguments{
\item{x}{An \code{fhx} instance.}

\item{spp}{Option to plot series with colors by species. A vector of species which corresponds to the series names given in \code{sppid}. Both \code{spp} and \code{sppid} need to be specified. Default plot gives no species colors.}

\item{sppid}{Option to plot series with colors by species. A vector of series names corresponding to species names given in \code{spp}. Every unique values in \code{x} series.names needs to have a corresponding species value. Both \code{spp} and \code{sppid} need to be specified. Default plot gives no species colors.}

\item{cluster}{Option to plot series with faceted by a factor. A vector of factors or characters which corresponds to the series names given in \code{clusterid}. Both \code{cluster} and \code{clusterid} need to be specified. Default plot is not faceted.}

\item{clusterid}{Option to plot series with faceted by a factor. A vector of series names corresponding to species names given in \code{cluster}. Every unique values in \code{x} series.names needs to have a corresponding cluster value. Both \code{cluster} and \code{clusterid} need to be specified.  Default plot is not faceted.}

\item{facet_type}{Type of \code{ggplot2} facet to use, if faceting. Must be either "grid" or "wrap". Default is "grid".}

\item{ylabels}{Optional boolean to remove y-axis (series name) labels and tick  marks. Default is TRUE.}

\item{yearlims}{Option to limit the plot to a range of years. This is a vector with two integers. The first integer gives the lower year for the range while the second integer gives the upper year. The default is to plot the full range of data given by \code{x}.}

\item{composite_rug}{A boolean option to plot a rug on the bottom of the plot. Default is FALSE.}

\item{filter_prop}{An optional argument if the user chooses to include a rug in their plot. This is passed to \code{composite}. See this function for details.}

\item{filter_min}{An optional argument if the user chooses to include a rug in their plot. This is passed to \code{composite}. See this function for details.}

\item{legend}{A boolean option allowing the user to choose whether a legend is included in the plot or not. Default is FALSE.}

\item{event_size}{An optional numeric vector that adjusts the size of fire event symbols on the plot. Default is \code{c("Scar" = 4, "Injury" = 2, "Pith/Bark" = 1.5)}.}

\item{rugbuffer_size}{An optional integer. If the user plots a rug, thiscontrols the amount of buffer whitespace along the y-axis between the rug and the main plot. Must be >= 2.}

\item{rugdivide_pos}{Optional integer if plotting a rug. Adjust the placement of the rug divider along the y-axis. Default is 2.}
}
\value{
A ggplot object for plotting or manipulation.
}
\description{
Create an ggplot2 object for plotting.
}

