% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_recording_count}
\alias{get_recording_count}
\title{Count the number of recording series for each year.}
\usage{
get_recording_count(x, injury_event = FALSE)
}
\arguments{
\item{x}{An fhx object.}

\item{injury_event}{Boolean indicating whether injuries should be considered event. Default is FALSE.}
}
\value{
A dataframe with a columns giving the year and corresponding number of recording events for that year.
}
\description{
Count the number of recording series for each year.
}

