% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_series}
\alias{get_series}
\title{Extract fhx observations for given series.}
\usage{
get_series(x, s)
}
\arguments{
\item{x}{An fhx object.}

\item{s}{Character vector of years you would like extracted from x.}
}
\value{
A dataframe with extracted observations.
}
\description{
Extract fhx observations for given series.
}

