% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea.R
\name{is.sea}
\alias{is.sea}
\title{Check if object is sea.}
\usage{
is.sea(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Boolean indicating whether `x` is an sea object.
}
\description{
Check if object is sea.
}
