% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_fhx}
\alias{is_fhx}
\title{Check if object is \code{fhx}.}
\usage{
is_fhx(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Boolean indicating whether \code{x} is an \code{fhx} object.
}
\description{
Check if object is \code{fhx}.
}
\examples{
data(lgr2)
is_fhx(lgr2)

}
\seealso{
\itemize{
\item \code{\link[=fhx]{fhx()}} constructs an \code{fhx} object.
\item \code{\link[=as_fhx]{as_fhx()}} casts data frame-like object into an \code{fhx} object.
\item \code{\link[=+.fhx]{+.fhx()}} concatenate multiple \code{fhx} objects together.
}
}
