% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{is_intervals}
\alias{is_intervals}
\title{Check if object is fire \code{intervals}}
\usage{
is_intervals(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Boolean indicating whether x is an \code{intervals} object.
}
\description{
Check if object is fire \code{intervals}
}
\seealso{
\code{\link[=intervals]{intervals()}} creates an \code{intervals} object.
}
