% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea.R
\name{is_sea}
\alias{is_sea}
\title{Check if object is \code{sea}}
\usage{
is_sea(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Boolean indicating whether \code{x} is a \code{sea} object.
}
\description{
Check if object is \code{sea}
}
\seealso{
\code{\link[=sea]{sea()}} creates a \code{sea} object.
}
