% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{last_year}
\alias{last_year}
\title{Last (most recent) year of an fhx series.}
\usage{
last_year(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
The maximum or last year of series in 'x'. 'NA' will be returned if 'NA' is in x$year.
}
\description{
Last (most recent) year of an fhx series.
}
