% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lgr2}
\alias{lgr2}
\title{Los Griegos Peak plot2 fire-history data}
\format{An fhx object with 26 series from 1366 to 2012 CE.}
\usage{
lgr2
}
\description{
An \code{fhx} object with fire-history data from Los Griegos Peak, New Mexico.
}
\seealso{
\link{lgr2_meta} Los Griegos Peak metadata.
}
\keyword{datasets}
