% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lgr2_meta}
\alias{lgr2_meta}
\title{Metadata for the Los Griegos Peak fire-history dataset}
\format{A data frame with 26 rows and 2 variables:
\itemize{
  \item TreeID: name of tree series
  \item SpeciesID: abbreviated tree species
}}
\usage{
lgr2_meta
}
\description{
A dataset with species information for the Los Griegos Peak plot2 fire-history dataset (`lgr2`).
}
\keyword{datasets}
