% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_filestrings}
\alias{list_filestrings}
\title{List of character strings to write to FHX file.}
\usage{
list_filestrings(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
A list with four members containing vectors: "head_line", 
    "subhead_line", "series_heading", and "body". Each referring 
    to a portion of an FHX file that the strings are dumped into.
}
\description{
List of character strings to write to FHX file.
}
\seealso{
write_fhx
}
