% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeResponsesList.R
\name{makeResponseList}
\alias{makeResponseList}
\title{Make list of responses for a Census question}
\usage{
makeResponseList(
  questionType = c("selectOne", "selectMultiple", "selectScale", "textBoxNumeric",
    "textBoxCharacter"),
  responses = NULL,
  reportingValues = NULL,
  scaleOptions = NULL,
  rname = NULL
)
}
\arguments{
\item{questionType}{Type of the question. One of "selectOne",
"selectMultiple", "selectScale", "textBoxNumeric",
textBoxCharacter".}

\item{responses}{Question responses. For "selectOne" and "selectMultiple"
questions, the text for possible response selections
as they were provided in the question for that year. For
"selectScale", the text for each sub-question. For
"textBox..." questions this should be NULL.}

\item{reportingValues}{Values used to refer to or store the responses in the
raw data outputs from Alchemer or other survey tools
used in prior years. For "selectScale" questions this
is the values used to store the scale options. For
"textBox..." questions this should be NULL.}

\item{scaleOptions}{For "selectScale" questions, the text or numeric values
for scale options as they were shown on the survey.}

\item{rname}{Response name. For "selectOne" and "selectMutliple" questions,
the text used to reference the response or sub-question in the
cleaned and weighted data file. For "textBox..." question this
should be NULL.}
}
\value{
A data.frame or list of possible responses to a Census question
}
\description{
Make list of responses for a Census question
}
