% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWriteInList.R
\name{makeWriteInList}
\alias{makeWriteInList}
\title{Make a list of write in responses for a Census question}
\usage{
makeWriteInList(writeInResponses, writeInTypes, wname)
}
\arguments{
\item{writeInResponses}{Wording for responses that have a write-in field}

\item{writeInTypes}{Type of write-in response, either "character" or "numeric"}

\item{wname}{wname for write-in responses.}
}
\value{
list of write-in response metadata.
}
\description{
Make a list of write in responses for a Census question
}
