% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{man}
\alias{man}
\title{Metadata And Notes (MAN)}
\usage{
man(
  qname,
  year,
  dname,
  questionText,
  questionType,
  responses = NULL,
  rname = NULL,
  scaleOptions = NULL,
  reportingValues = NULL,
  writeInResponses = NULL,
  writeInTypes = NULL,
  wname = NULL,
  notes.qname = NA,
  notes.year = NA
)
}
\arguments{
\item{qname}{Question name. The name used to reference a question for
analysis. This name should be the same across years to enable
trend analysis.}

\item{year}{The year the question was asked. Integer value with the format
19## or 20##.}

\item{dname}{Data file name. The name the question is stored under in the
cleaned and weighted data file for Census responses for the
given year. For questions of type 'selectMultiple' prior to
2019, this value will be unique for each option. For years 2019
and later, this will be the same as qname.}

\item{questionText}{The full text of the question as it was asked.}

\item{questionType}{Type of the question. One of "selectOne",
"selectMultiple", "selectScale", "textBoxNumeric",
textBoxCharacter".}

\item{responses}{Question responses. For "selectOne" and "selectMultiple"
questions, the text for possible response selections
as they were provided in the question for that year. For
"selectScale", the text for each sub-question. For
"textBox..." questions this should be NULL.}

\item{rname}{Response name. For "selectOne" and "selectMutliple" questions,
the text used to reference the response or sub-question in the
cleaned and weighted data file. For "textBox..." question this
should be NULL.}

\item{scaleOptions}{For "selectScale" questions, the text or numeric values
for scale options as they were shown on the survey.}

\item{reportingValues}{Values used to refer to or store the responses in the
raw data outputs from Alchemer or other survey tools
used in prior years. For "selectScale" questions this
is the values used to store the scale options. For
"textBox..." questions this should be NULL.}

\item{writeInResponses}{Character vector of responses that have a write-in
option. This should be a subset of fields provided
in responses. This should be NULL for "textBox..."
questions.}

\item{writeInTypes}{A character vector of the same length as writeInRespones.
Values specify the type of write-in question, either "character" or "numeric".}

\item{wname}{Write-in name. Name used to refer to write-in responses in
cleaned and weighted data files.}

\item{notes.qname}{Notes for a question that are pertinent to all years. This
will be combined with the same field across all years
when man objects are concatenated.}

\item{notes.year}{Notes for a question that are pertinent to the given year.}
}
\value{
A 'man' object.
}
\description{
The function man() creates a metadata and notes (man) object
which describes the format of a question from the Black Rock City Census.
}
\examples{
# 'selectOne' type question no write-in
virgin25 <- man(qname = "virgin",
                year = 2025,
                dname = "virgin",
                questionText = "Was this your first time visiting Black Rock City?",
                questionType = "selectOne",
                responses = c("Yes, it was my first time",
                              "No, I have been to Black Rock City before",
                              "I did not go to Black Rock City in 2025"),
                rname = c("virgin",
                          "notVirgin",
                          "didNotGoIn2025"),
                reportingValues = c(1, 2, 3))

# 'selectMany' type question with write-in
raceEth25 <- man(qname = "raceEth",
                 year = 2025,
                 dname = "raceEth",
                 questionText = "Which category best describes your race or ethnicity?",
                 questionType = "selectMultiple",
                 responses = c("Native American or Alaska Native",
                               "Asian",
                               "Black or African American",
                               "Native Hawaiian or other Pacific Islander",
                               "White/Caucasian",
                               "Middle Eastern or North African",
                               "Hispanic/Latino (e.g., Mexican, Puerto Rican, Cuban, Colombian)",
                               "Other (please specify):"),
                 rname = c("nativeAmericanAlaskan",
                           "asian",
                           "black",
                           "nativeHawaiianPacificIslander",
                           "white",
                           "middleEasternNorthAfrican",
                           "hispanic",
                           "other"),
                reportingValues = c(2, 1, 5, 8, 4, 7, 4, 6),
                writeInResponses = c("Other (please specify):"),
                wname = "other",
                writeInTypes = c("character"),
                notes.qname = "The options for this question are based on
                               the recommendations of the US Census.  The
                               options were updated in 2022.")

# 'selectScale' type question
mediaManagedByBMP25 <- man(qname = "mediaManagedByBMP",
                           year = 2025,
                           dname = "mediaManagedByBMP",
                           questionText = "Which of the following media managed by Burning Man Project do you use to receive news, information, and events related to Burning Man, and to engage in discussion on Burning Man topics?",
                           questionType = "selectScale",
                           responses = c("Jack Rabbit Speaks (JRS) e-newsletter",
                                         "Burning Man website",
                                         "Burning Man Journal",
                                         "Burning Man Hive",
                                         "Social media managed by Burning Man Project (Facebook, Instagram, Twitter/X, YouTube, etc.)",
                                         "Eplaya"),
                           scaleOptions = c("Never", "Rarely", "Often"),
                           reportingValues = c(1, 2, 3),
                           rname = c("jackrabbitSpeaks",
                                     "burningManWebsite",
                                     "burningManJournal",
                                     "burningManHive",
                                     "socialMedia",
                                     "ePlaya"))

# Pre-alchemer (2022) example
virgin19 <- man(qname = "virgin",
                year = 2019,
                dname = "virgin",
                questionText = "Was this your first time visiting Black Rock City?",
                questionType = "selectOne",
                responses = c("Yes, it was my first time",
                              "No, I had been to Black Rock City before",
                              "I did not go to Black Rock City in 2019"),
                rname = c("virgin",
                          "notVirgin",
                          "didNotGoIn2019"),
                reportingValues = c(1, 2, 3))

# Concatenate man objects (are many mans a men?)
virginMan <- virgin25 + virgin19

}
