% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{max.intervals}
\alias{max.intervals}
\title{Maximum interval.}
\usage{
\method{max}{intervals}(x, ...)
}
\arguments{
\item{x}{An intervals object.}

\item{...}{Additional arguments passed to \code{max}.}
}
\value{
Numeric or NA.
}
\description{
Maximum interval.
}
