% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{mean.intervals}
\alias{mean.intervals}
\title{Interval arithmetic mean.}
\usage{
\method{mean}{intervals}(x, ...)
}
\arguments{
\item{x}{An intervals object.}

\item{...}{Additional arguments passed to \code{mean}.}
}
\value{
Numeric or NA.
}
\description{
Interval arithmetic mean.
}
