% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean_interval}
\alias{mean_interval}
\title{Calculate mean fire interval of an fhx series.}
\usage{
mean_interval(x, injury_event = FALSE)
}
\arguments{
\item{x}{An fhx object.}

\item{injury_event}{Boolean indicating whether injuries should be considered event.}
}
\value{
The mean fire interval observed in the series.
}
\description{
Calculate mean fire interval of an fhx series.
}
