% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{median.intervals}
\alias{median.intervals}
\title{Interval median.}
\usage{
\method{median}{intervals}(x, ...)
}
\arguments{
\item{x}{An intervals object.}

\item{...}{Additional arguments passed to \code{median}.}
}
\value{
Numeric or NA.
}
\description{
Interval median.
}
