% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehistory.R
\name{order.fhx}
\alias{order.fhx}
\title{Reorder the series names of an fhx instance.}
\usage{
order.fhx(x)
}
\arguments{
\item{x}{An fhx instance to be reorder.}
}
\value{
A copy of \code{x} with reordered series.
}
\description{
Reorder the series names of an fhx instance.
}

