% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{percent_scarred}
\alias{percent_scarred}
\title{Percent scarred time series}
\usage{
percent_scarred(x, injury_event = FALSE)
}
\arguments{
\item{x}{An fhx object.}

\item{injury_event}{Boolean indicating whether years with injury events should be considered as scars. Defaults to FALSE.}
}
\value{
A data.frame with four columns: \code{Year}, \code{NumRec} with the number of recording trees,
\code{NumScars} with the number of fire scars and/or events, and \code{PercScarred} with the proportion of scars/events.
}
\description{
Percent scarred time series
}
\examples{
data("pgm")
percent_scarred(pgm)

}
