% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehistory.R
\name{plot.fhx}
\alias{plot.fhx}
\title{Plot an fhx object.}
\usage{
\method{plot}{fhx}(...)
}
\arguments{
\item{...}{Arguments passed on to \code{ggplot.fhx}.}
}
\description{
Plot an fhx object.
}

