% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{plot.intervals}
\alias{plot.intervals}
\title{Plot an intervals object.}
\usage{
\method{plot}{intervals}(...)
}
\arguments{
\item{...}{Arguments passed on to \code{plot_intervals_dist}.}
}
\description{
Plot an intervals object.
}
\examples{
data(pgm)
interv <- intervals(composite(pgm))

plot(interv, binwidth = 5)
}
