% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_season}
\alias{plot_season}
\title{Create an ggplot2 object for plotting fhx event seasonality.}
\usage{
plot_season(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
A ggplot object for plotting or manipulation.
}
\description{
Create an ggplot2 object for plotting fhx event seasonality.
}
\examples{
data(lgr2)
plot_season(lgr2)

}

