% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehistory.R
\name{+.fhx}
\alias{+.fhx}
\title{Concatenate two fhx instance.}
\usage{
\method{+}{fhx}(a, b)
}
\arguments{
\item{a}{An fhx instance.}

\item{b}{The fhx instance to be appended.}
}
\value{
An fhx instance with the information from \code{a} and \code{b}. Duplicates are resolved with \code{fire::resolve_duplicates()}.
}
\description{
Concatenate two fhx instance.
}

