% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_man.R
\name{+.man}
\alias{+.man}
\title{Concatenate man objects}
\usage{
\method{+}{man}(man1, man2)
}
\arguments{
\item{man1}{An object of class 'man'.}

\item{man2}{An object of class 'man'.}
}
\value{
An object of class 'man'.
}
\description{
Concatenate man objects
}
