% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pme}
\alias{pme}
\title{Pajarito Mountain East fire-history data}
\format{
An \code{fhx} object with 17 series from 1702 to 1993 CE.
}
\source{
https://www1.ncdc.noaa.gov/pub/data/paleo/firehistory/firescar/northamerica/uspme001.fhx
}
\usage{
pme
}
\description{
An \code{fhx} object with fire-history data.
}
\keyword{datasets}
