% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmr}
\alias{pmr}
\title{Pajarito Mountain Ridge fire-history data}
\format{
An \code{fhx} object with 23 series from 1626 to 1993 CE.
}
\source{
https://www1.ncdc.noaa.gov/pub/data/paleo/firehistory/firescar/northamerica/uspmr001.fhx
}
\usage{
pmr
}
\description{
An \code{fhx} object with fire-history data.
}
\keyword{datasets}
