% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmw}
\alias{pmw}
\title{Pajarito Mountain West fire-history data}
\format{
An \code{fhx} object with 11 series from 1617 to 1993 CE.
}
\source{
https://www1.ncdc.noaa.gov/pub/data/paleo/firehistory/firescar/northamerica/uspmw001.fhx
}
\usage{
pmw
}
\description{
An \code{fhx} object with fire-history data.
}
\keyword{datasets}
