% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.fhx}
\alias{print.fhx}
\title{Print an \code{fhx} object.}
\usage{
\method{print}{fhx}(x, ...)
}
\arguments{
\item{x}{An \code{fhx} object.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\description{
Print an \code{fhx} object.
}
\examples{
data(lgr2)
print(lgr2)

}

