% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea.R
\name{print.sea}
\alias{print.sea}
\title{Print an sea objects.}
\usage{
\method{print}{sea}(x, ...)
}
\arguments{
\item{x}{An intervals object.}

\item{...}{Additional arguments that are tossed.}
}
\description{
Print an sea objects.
}
