% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{quantile.intervals}
\alias{quantile.intervals}
\title{Fit distribution quantiles.}
\usage{
\method{quantile}{intervals}(x, q = c(0.125, 0.5, 0.875), ...)
}
\arguments{
\item{x}{An intervals object.}

\item{q}{Vector giving the desired quantiles.}

\item{...}{Additional arguments passed to the quantile function of the fit distribution.}
}
\description{
Fit distribution quantiles.
}
\examples{
data(pgm)
intervs <- intervals(composite(pgm))
quantile(intervs)

# Or you can pass in your own quantiles:
quantile(intervs, q = c(0.25, 0.5, 0.75))

}
