% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{quantile}
\alias{quantile}
\title{Sample Quantiles.}
\usage{
quantile(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Sample Quantiles.
}
