% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{read.fhx}
\alias{read.fhx}
\title{Read input FHX file body from \code{fname} and use to return an \code{fhx} object.}
\usage{
read.fhx(fname, encoding = getOption("encoding"))
}
\arguments{
\item{fname}{Name of target FHX file. Needs to be in format version 2.}

\item{encoding}{Encoding to use when reading the FHX file. The default is to use the system.}
}
\value{
An \code{fhx} object.
}
\description{
Read input FHX file body from \code{fname} and use to return an \code{fhx} object.
}

