% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_census.R
\name{read_census}
\alias{read_census}
\title{Read in Census data}
\usage{
read_census(filePaths, year, weights, metaData)
}
\arguments{
\item{filePaths}{character vector of file paths pointing to Census tsv files}

\item{weights}{character vector with the same length as year corresponding
to the name of the weight variable used for each year.}

\item{metaData}{A census.meta object.  Only questions included in the
census.meta object will be included in the census.data object.}

\item{years}{numeric vector with the same length as filePaths of years being
read into the census.data object and corresponding to each file path.}
}
\value{
A census.data object.
}
\description{
Read in Census data
}
\examples{
# example code
}
