% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{rec_type2abvr}
\alias{rec_type2abvr}
\title{Convert rec_type char to abreviated fhx file event char.}
\usage{
rec_type2abvr(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string.
}
\description{
Convert rec_type char to abreviated fhx file event char.
}
