% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recording_finder}
\alias{recording_finder}
\title{Subset `rings` data.frame to years that are considered recording.}
\usage{
recording_finder(x, injury_event)
}
\arguments{
\item{x}{A dataframe from an fhx object.}

\item{injury_event}{Boolean indicating whether injuries should be considered event.}
}
\value{
A dataframe with a column of each year which is 'recording'.
}
\description{
Subset `rings` data.frame to years that are considered recording.
}

