% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_duplicate_series_end}
\alias{remove_duplicate_series_end}
\title{Remove duplicate end-years in FHX objects}
\usage{
remove_duplicate_series_end(x, s, keep_event = TRUE)
}
\arguments{
\item{x}{An \code{fhx} object}

\item{s}{One or more series to correct. If left blank, the function will find
and fix all series. Use this if you would like to specify which outer ring
type to keep for different series}

\item{keep_event}{Boolean, defaults to TRUE to keep the event codes (scars or
injuries). FALSE will select the inner/outer ring codes instead}
}
\value{
A revised \code{fhx} object with corrected series end years. Check your
work via \link{check_series}
}
\description{
FHX files will not handle duplicate rec_types in a single year. Often the
duplication occurs at the series ends -- either the beginning or end -- where
one might note a scar or injury that starts or ends the series. To generate
an FHX file, one of the duplicate rec_types must be removed, or its year
changed. This function removes one of the duplicate rec_types to allow for
FHX files to be written via \code{\link[=write_fhx]{write_fhx()}}.
}
