% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehistory.R
\name{resolve_duplicates}
\alias{resolve_duplicates}
\title{Merge/remove duplicate observations in an fhx object.}
\usage{
resolve_duplicates(x)
}
\arguments{
\item{x}{An fhx instance.}
}
\value{
A copy of \code{x} with duplicates removed.
}
\description{
Merge/remove duplicate observations in an fhx object.
}

