% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehistory.R
\name{rug.filter}
\alias{rug.filter}
\title{Filter fire events in `x` returning years with prominent fires.}
\usage{
rug.filter(x, filter.prop = 0.25, filter.min = 2)
}
\arguments{
\item{filter.prop}{The proportion of fire events to recording series needed in order to be considered. Default is 0.25.}

\item{filter.min}{The minimum number of recording series needed to be considered a fire event. Default is 2 recording series.}
}
\value{
A vector of years from `x`.
}
\description{
Filter fire events in `x` returning years with prominent fires.
}

