% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea.R
\name{run_sea}
\alias{run_sea}
\title{Perform superposed epoch analysis.}
\usage{
run_sea(x, key, years_before = 6, years_after = 4, key_period = TRUE,
  n_iter = 1000)
}
\description{
Perform superposed epoch analysis.
}
\details{
This function, 'run_sea', is a development version of burnr::sea and has benn depreciated. Please use \code{sea}.
}
