% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{s_count_fire}
\alias{s_count_fire}
\title{Number of fire events in an fhx series.}
\usage{
s_count_fire(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
The number of fire events observed in the series.
}
\description{
Number of fire events in an fhx series.
}

