% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{s_count_year}
\alias{s_count_year}
\title{Number of years of an fhx series.}
\usage{
s_count_year(x)
}
\arguments{
\item{x}{An fhx object's 'rings' data.frame.}
}
\value{
The difference between the first and last observations in the series. 'NA' will be returned if 'NA' is in 'x$year'.
}
\description{
Number of years of an fhx series.
}

