% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{s_first_year}
\alias{s_first_year}
\title{First (earliest) year of an fhx series.}
\usage{
s_first_year(x)
}
\arguments{
\item{x}{An fhx object's 'rings' data.frame.}
}
\value{
The minimum or first year of series in 'x'.
}
\description{
First (earliest) year of an fhx series.
}

