% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{s_inner_type}
\alias{s_inner_type}
\title{Type of observation in the first (earliest) year of an fhx series.}
\usage{
s_inner_type(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
The a factor giving the type of observation in the first observation of the series.
}
\description{
Type of observation in the first (earliest) year of an fhx series.
}

