% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{s_outer_type}
\alias{s_outer_type}
\title{Type of observation in the last (most recent) year of an fhx series.}
\usage{
s_outer_type(x)
}
\arguments{
\item{x}{An fhx object's 'rings' data.frame.}
}
\value{
The a factor giving the type of observation in the last observation of the series.
}
\description{
Type of observation in the last (most recent) year of an fhx series.
}

