% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{samp_depth}
\alias{samp_depth}
\title{Calculate the sample depth of an fhx object}
\usage{
samp_depth(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
A data.frame containing the years and number of trees
}
\description{
Calculate the sample depth of an fhx object
}

