% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{sd.intervals}
\alias{sd.intervals}
\title{Interval standard deviation.}
\usage{
\method{sd}{intervals}(x, ...)
}
\arguments{
\item{x}{An intervals object.}

\item{...}{Additional arguments passed to \code{sd}.}
}
\value{
Numeric or NA.
}
\description{
Interval standard deviation.
}
