% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{series_stats}
\alias{series_stats}
\title{A function to generate series-level descriptive statistics}
\usage{
series_stats(x)
}
\arguments{
\item{x}{An fhx instance}
}
\value{
A data.frame containing series-level statistics
}
\description{
A function to generate series-level descriptive statistics
}

