% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{sort.fhx}
\alias{sort.fhx}
\title{Resort the series names of an fhx instance.}
\usage{
\method{sort}{fhx}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{An fhx instance to be reorder.}

\item{decreasing}{Logical. Decreasing sorting? Defaults to FALSE.}

\item{...}{Additional arguments that fall off the face of the universe.}
}
\value{
A copy of \code{x} with reordered series.
}
\description{
Resort the series names of an fhx instance.
}

