% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset.fhx}
\alias{subset.fhx}
\title{Subset fhx observations for given years or series.}
\usage{
\method{subset}{fhx}(x, subset, ...)
}
\arguments{
\item{x}{An fhx object.}

\item{subset}{Character vector of series, or integer vector of years to extract from x.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
An fhx object with extracted observations.
}
\description{
Subset fhx observations for given years or series.
}
\examples{
data(lgr2)
plot(subset(lgr2, 'LGR46'))

plot(subset(lgr2, c('LGR41', 'LGR46')))

plot(subset(lgr2, 1550:2020))

}

