% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary.fhx}
\alias{summary.fhx}
\title{Summary of \code{fhx} object}
\usage{
\method{summary}{fhx}(object, ...)
}
\arguments{
\item{object}{An \code{fhx} object.}

\item{...}{Additional arguments that are tossed out.}
}
\value{
A \code{summary.fhx} object.
}
\description{
Summary of \code{fhx} object
}
\seealso{
\code{\link[=series_stats]{series_stats()}} basic statistics for series in an \code{fhx} object.
}
