% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree_stats}
\alias{tree_stats}
\title{A function to generate tree-level descriptive statics}
\usage{
tree_stats(x)
}
\arguments{
\item{x}{An fhx instance}
}
\value{
A data.frame containing tree-level statistics
}
\description{
A function to generate tree-level descriptive statics
}

