% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{year_range}
\alias{year_range}
\title{Range of years for \code{fhx} object.}
\usage{
year_range(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
An integer vector or \code{NULL}.
}
\description{
Range of years for \code{fhx} object.
}
\examples{
data(lgr2)
year_range(lgr2)

}
