% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yearly_recording}
\alias{yearly_recording}
\title{Count the number of recording series for each year in an \code{fhx} object}
\usage{
yearly_recording(x, injury_event = FALSE)
}
\arguments{
\item{x}{An \code{fhx} object.}

\item{injury_event}{Boolean indicating whether injuries should be considered
events. Default is \code{FALSE}.}
}
\value{
A data frame with columns giving the year and recording events count.
}
\description{
Count the number of recording series for each year in an \code{fhx} object
}
\examples{
data(lgr2)
yearly_recording(lgr2)

}
