% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.fhx}
\alias{as.fhx}
\title{Convert to fhx object.}
\usage{
as.fhx(x)
}
\arguments{
\item{x}{A data frame or list-like object. Must have named elements or columns for "year", "series", and "rec_type".}
}
\value{
`x` cast to an fhx object.
}
\description{
Convert to fhx object.
}
\details{
The "year", "series", and "rec_type" in \code{x} will be pass through \code{as.numeric}, \code{as.factor}, and \code{burnr::make_rec_type} before being passed to \code{burnr::fhx}.
}
\examples{
data(lgr2)
example_dataframe <- as.data.frame(lgr2)
back_to_fhx <- as.fhx(example_dataframe)

}
