% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_scar}
\alias{count_scar}
\title{Number of scar events in an fhx series.}
\usage{
count_scar(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
The number of fire events observed in the series.
}
\description{
Number of scar events in an fhx series.
}
