% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete}
\alias{delete}
\title{Remove series or years from an fhx object.}
\usage{
delete(x, s, yr)
}
\arguments{
\item{x}{An fhx object.}

\item{s}{Character vector of series to erase from x.}

\item{yr}{Integer vector of years to erase from x.}
}
\value{
An fhx object with observations erased.
}
\description{
Remove series or years from an fhx object.
}
\details{
You can combine s and yr to specify years within select series to remove.
}
\examples{
data(lgr2)
plot(delete(lgr2, s = 'LGR46'))

plot(delete(lgr2, yr = 1300:1550))

}
