% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_ggplot}
\alias{get_ggplot}
\title{Create a ggplot2 object for plotting.}
\usage{
get_ggplot(...)
}
\arguments{
\item{...}{Arguments passed on to \code{plot_demograph}.}
}
\description{
This function is depreciated. Please use `plot_demograph()`.
}

