% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_series}
\alias{get_series}
\title{Extract fhx observations for given series.}
\usage{
get_series(x, s)
}
\arguments{
\item{x}{An fhx object.}

\item{s}{Character vector of series you would like extracted from x.}
}
\value{
A dataframe with extracted observations.
}
\description{
Extract fhx observations for given series.
}
\examples{
data(lgr2)
get_series(lgr2, 'LGR46')

get_series(lgr2, c('LGR41', 'LGR46'))

}
