% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_year}
\alias{get_year}
\title{Extract fhx observations for given years.}
\usage{
get_year(x, yr)
}
\arguments{
\item{x}{An fhx object.}

\item{yr}{Integer vector of year(s) you would like extracted from x.}
}
\value{
A dataframe with extracted observations.
}
\description{
Extract fhx observations for given years.
}
\examples{
data(lgr2)
get_year(lgr2, 1806)

get_year(lgr2, 1805:1807)

}
