% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{Constructor for S3 intervals class.}
\usage{
intervals(comp, densfun = "weibull", ...)
}
\arguments{
\item{comp}{A composite fhx instance.}

\item{densfun}{String giving desired distribution to fit. Can be "weibull" or "lognormal". Default is "weibull".}

\item{...}{Arguments passed to \code{fit_distr}.}
}
\value{
An intervals instance.
}
\description{
Constructor for S3 intervals class.
}
