% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.fhx}
\alias{is.fhx}
\title{Check if object is fhx.}
\usage{
is.fhx(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Boolean indicating whether `x` is an fhx object.
}
\description{
Check if object is fhx.
}
\examples{
data(lgr2)
is.fhx(lgr2)

}
