% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{is.intervals}
\alias{is.intervals}
\title{Check if object is intervals.}
\usage{
is.intervals(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Boolean indicating whether `x` is an intervals object.
}
\description{
Check if object is intervals.
}
