% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lgr2_meta}
\alias{lgr2_meta}
\title{Metadata for the Los Griegos Peak fire-history dataset}
\format{
A \code{data.frame} with 26 rows and 2 variables:
\itemize{
\item "TreeID": Name of tree series.
\item "SpeciesID": Abbreviated tree species
}
}
\usage{
lgr2_meta
}
\description{
A data frame with species information for the Los Griegos Peak plot2
fire-history dataset (\link{lgr2}).
}
\seealso{
\link{lgr2} Log Griegos Peak fire-history data.
}
\keyword{datasets}
