% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rec_type}
\alias{make_rec_type}
\title{Turn character vector into factor with proper fhx levels.}
\usage{
make_rec_type(x)
}
\arguments{
\item{x}{A character vector containing one or more rec_type-like strings.}
}
\value{
A factor with appropriate fhx levels.
}
\description{
Turn character vector into factor with proper fhx levels.
}
\examples{
make_rec_type('null_year')

make_rec_type(c('null_year', 'late_fs'))

}
