% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{min.intervals}
\alias{min.intervals}
\title{Minimum interval.}
\usage{
\method{min}{intervals}(x, ...)
}
\arguments{
\item{x}{An intervals object.}

\item{...}{Additional arguments passed to \code{min}.}
}
\value{
Numeric or NA.
}
\description{
Minimum interval.
}
