% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgm}
\alias{pgm}
\title{Peggy Mesa fire-history data}
\format{
An \code{fhx} object with 41 series from 1555 to 2013 CE.
}
\source{
Guiterman, Christopher H., Ellis Q. Margolis, and
Thomas W. Swetnam. 2015. "Dendroecological Methods For Reconstructing
High-Severity Fire In Pine-Oak Forests." Tree-Ring Research 71 (2): 67-77.
doi:10.3959/1536-1098-71.2.67.
}
\usage{
pgm
}
\description{
An \code{fhx} object with fire-history data from Peggy Mesa.
}
\seealso{
\itemize{
\item \link{pgm_meta} Peggy Mesa metadata.
\item \link{pgm_pdsi} PDSI time-series for Peggy Mesa site.
}
}
\keyword{datasets}
