% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgm_pdsi}
\alias{pgm_pdsi}
\title{Reconstructed PDSI time series for the Peggy Mesa fire-history dataset}
\format{A data frame with 2004 rows and 1 variables. Row names give the year for the reconstructed value:
\itemize{
  \item RECON: The reconstructed PDSI series.
}}
\usage{
pgm_pdsi
}
\description{
A tree-ring reconstructed Palmer Drought-Severity Index time series corresponding to the Peggy Mesa fire-history dataset (`pgm`) -- specifically, the Jemez Mountains area (gridpoint 133). The reconstruction is from The North American Drought Atlas (Cook and Krusic 2004).
}
\keyword{datasets}
