% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{plot_intervals_dist}
\alias{plot_intervals_dist}
\title{Basic intervals distribution plot.}
\usage{
plot_intervals_dist(x, binwidth = NULL)
}
\arguments{
\item{x}{An intervals object.}

\item{binwidth}{A sea object.}
}
\value{
A ggplot object.
}
\description{
Basic intervals distribution plot.
}
