% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{+.fhx}
\alias{+.fhx}
\title{Concatenate or combine two fhx objects.}
\usage{
\method{+}{fhx}(a, b)
}
\arguments{
\item{a}{An fhx object.}

\item{b}{The fhx object to be append.}
}
\value{
An fhx object with the series from \code{a} and \code{b}.
}
\description{
Concatenate or combine two fhx objects.
}
\examples{
data(lgr2)
data(pgm)
plot(lgr2 + pgm)

}
