% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{print.intervals}
\alias{print.intervals}
\title{Print an intervals objects.}
\usage{
\method{print}{intervals}(x, ...)
}
\arguments{
\item{x}{An intervals object.}

\item{...}{Additional arguments that are tossed.}
}
\description{
Print an intervals objects.
}
