% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_fhx}
\alias{read_fhx}
\title{Read FHX2 file and return an \code{fhx} object.}
\usage{
read_fhx(fname, encoding, text)
}
\arguments{
\item{fname}{Name of target FHX file. Needs to be in format version 2.}

\item{encoding}{Encoding to use when reading the FHX file. The default is to use the system.}

\item{text}{Character string. If \code{fname} is not provided and 
\code{text} is, then data is read from \code{text} using a text connection.}
}
\value{
An \code{fhx} object.
}
\description{
Read FHX2 file and return an \code{fhx} object.
}
\examples{
\dontrun{
d <- read_fhx('afile.fhx')
}

}
