% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_duplicates}
\alias{resolve_duplicates}
\title{Test for duplicate observations in an fhx object.}
\usage{
resolve_duplicates(x)
}
\arguments{
\item{x}{An fhx instance.}
}
\value{
A copy of \code{x} with duplicates removed.
}
\description{
Test for duplicate observations in an fhx object.
}

