% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sample_depth}
\alias{sample_depth}
\title{Calculate the sample depth of an \code{fhx} object}
\usage{
sample_depth(x)
}
\arguments{
\item{x}{An \code{fhx} object.}
}
\value{
A data frame containing the years and number of observations.
}
\description{
Calculate the sample depth of an \code{fhx} object
}
\seealso{
\code{\link[=series_stats]{series_stats()}} basic statistics for series in an \code{fhx} object.
}
