% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sample_depth}
\alias{sample_depth}
\title{Calculate the sample depth of an fhx object}
\usage{
sample_depth(a)
}
\arguments{
\item{a}{An fhx object.}
}
\value{
A data.frame containing the years and number of trees
}
\description{
Calculate the sample depth of an fhx object
}

