% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{series_mean_interval}
\alias{series_mean_interval}
\title{Calculate mean fire interval of a single fhx series.}
\usage{
series_mean_interval(x, injury_event = FALSE)
}
\arguments{
\item{x}{An fhx object with a single series. For proper fire intervals see `intervals()`.}

\item{injury_event}{Boolean indicating whether injuries should be considered event.}
}
\value{
The mean fire interval observed in the series.
}
\description{
Calculate mean fire interval of a single fhx series.
}
\seealso{
intervals()
}
