% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{series_names}
\alias{series_names}
\title{Get \code{fhx} series names.}
\usage{
series_names(x)
}
\arguments{
\item{x}{An fhx object.}
}
\value{
A character vector or \code{NULL}.
}
\description{
Get \code{fhx} series names.
}
\examples{
data(lgr2)
series_names(lgr2)

}

