% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary.fhx}
\alias{summary.fhx}
\title{Summary of `fhx` object}
\usage{
\method{summary}{fhx}(object, ...)
}
\arguments{
\item{object}{An fhx object.}

\item{...}{Additional arguments.}
}
\value{
A summary.fhx object.
}
\description{
Summary of `fhx` object
}
