% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_fhx}
\alias{write_fhx}
\title{Write an fhx object to a new FHX2 file.}
\usage{
write_fhx(x, fname = "")
}
\arguments{
\item{x}{An fhx object.}

\item{fname}{Output filename.}
}
\description{
Write an fhx object to a new FHX2 file.
}
\examples{
\dontrun{
data(lgr2)
write_fhx(lgr2, 'afile.fhx')
}

}
